Remove deprecated two line item for standard replacement
Fix issues with background service and how we were starting the iomanager

Make sure we shut down the service when the device is disconnected.

Only scan once as the app is auto launched on plugging the GPS in and on pause and resume.
Alternative is to just stop scanning once we launch the background service.